pyproject.toml
[tool.poetry]
name = "bookmarks"
version = "0.1.0"
description = ""
authors = ["fullonic <somnium@riseup.net>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.4"
django-extensions = "^3.1.0"
bs4 = "^0.0.1"
httpx = "^0.16.1"
watchdog = "^1.0.2"
pytest = "^6.2.1"
pytest-django = "^4.1.0"
djangorestframework = "^3.12.2"
djangorestframework-simplejwt = "^4.6.0"

[tool.poetry.dev-dependencies]
ipython = "^7.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
